For now don't use XPathFunctionRegistry to check for XPath function services
Only check for the existence of classes.

For now don't let Modules and Components depend on any Java services
except other Modules and Components (through ModuleRegistry). Java
services may depend on non-Java components, but not the other way
around. Otherwise we will get circular
dependencies. XPathFunctionRegistry may depend on function
implementations that (directly or indirectly) depend on a certain
Module that also binds XPathFunctionRegistry. Even if
XPathFunctionRegistry is made to deal with this, there will still be
issues when there are internal (within the same module) dependencies
between Java and non-Java components, because a Module is only made
available to ModuleRegistry when it has been fully initialized.

The additional benefit of not using XPathFunctionRegistry is that we
don't rely on Saxon specific packages.